nasty windows race condition
