+++ conda/utils.py
@@ -6,9 +6,7 @@ import hashlib
 import logging
 import os
 import re
-import subprocess
 import sys
-import tempfile
 import time
 from functools import partial
 from os.path import abspath, isdir, join, isfile
@@ -115,8 +113,7 @@ def try_write(dir_path, heavy=False):
         except (IOError, OSError):
             return False
         finally:
-            if isfile(temp_filename):
-                exp_backoff_fn(os.unlink, temp_filename)
+            exp_backoff_fn(lambda f: isfile(f) and os.unlink(f), temp_filename)
     else:
         return os.access(dir_path, os.W_OK)
 
@@ -148,32 +145,6 @@ def path_to_url(path):
 #     assert url.startswith('file://')
 #     url_parts = urlparse.urlparse(url)
 #     return abspath(join(url_parts.netloc, url_parts.path))
-
-
-def run_in(command, shell, cwd=None, env=None):
-    if hasattr(shell, "keys"):
-        shell = shell["exe"]
-    if shell == 'cmd.exe':
-        cmd_script = tempfile.NamedTemporaryFile(suffix='.bat', mode='wt', delete=False)
-        cmd_script.write(command)
-        cmd_script.close()
-        cmd_bits = [shells[shell]["exe"]] + shells[shell]["shell_args"] + [cmd_script.name]
-        try:
-            p = subprocess.Popen(cmd_bits, stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-                                 cwd=cwd, env=env)
-            stdout, stderr = p.communicate()
-        finally:
-            os.unlink(cmd_script.name)
-    elif shell == 'powershell':
-        raise NotImplementedError
-    else:
-        cmd_bits = ([shells[shell]["exe"]] + shells[shell]["shell_args"] +
-                    [translate_stream(command, shells[shell]["path_to"])])
-        p = subprocess.Popen(cmd_bits, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        stdout, stderr = p.communicate()
-    streams = [u"%s" % stream.decode('utf-8').replace('\r\n', '\n').rstrip("\n")
-               for stream in (stdout, stderr)]
-    return streams
 
 
 def path_identity(path):
